refactor: LottoService 분리
